index: Extract `ser` module
