+++ cargo-registry-index/lib.rs
@@ -6,9 +6,11 @@ extern crate tracing;
 mod credentials;
 mod data;
 mod repo;
+mod ser;
 #[cfg(feature = "testing")]
 pub mod testing;
 
 pub use crate::credentials::Credentials;
-pub use crate::data::{write_crates, Crate, Dependency, DependencyKind};
+pub use crate::data::{Crate, Dependency, DependencyKind};
 pub use crate::repo::{Repository, RepositoryConfig};
+pub use crate::ser::write_crates;
